Move substitution-related stuff into a separate module
Fix unit tests

add test
