+++ booster/library/Booster/Pattern/Match.hs
@@ -36,12 +36,12 @@ import Booster.Definition.Attributes.Base (KListDefinition, KMapDefinition, KSet
 import Booster.Definition.Base
 import Booster.Pattern.Base
 import Booster.Pattern.Pretty
+import Booster.Pattern.Substitution qualified as Substitution
 import Booster.Pattern.Util (
     checkSymbolIsAc,
     freeVariables,
     isConstructorSymbol,
     sortOfTerm,
-    substituteInTerm,
  )
 
 -- | Result of matching a pattern to a subject (a substitution or an indication of what went wrong)
@@ -665,7 +665,7 @@ matchMaps
             -- before matching map keys.
                 enqueueMapProblem (KMap def patKeyVals patRest) (KMap def subjKeyVals subjRest)
             else do
-                let patternKeyVals = map (first (substituteInTerm st.mSubstitution)) patKeyVals
+                let patternKeyVals = map (first (Substitution.substituteInTerm st.mSubstitution)) patKeyVals
                 -- check for duplicate keys
                 checkDuplicateKeys patternKeyVals patRest
                 checkDuplicateKeys subjKeyVals subjRest
@@ -782,13 +782,13 @@ bindVariable matchType var term@(Term termAttrs _) = do
         Nothing -> do
             let
                 -- apply existing substitutions to term
-                term' = substituteInTerm currentSubst term
+                term' = Substitution.substituteInTerm currentSubst term
             when (var `Set.member` freeVariables term') $
                 failWith (VariableRecursion var term)
             let
                 -- substitute in existing substitution terms
                 currentSubst' =
-                    Map.map (substituteInTerm $ Map.singleton var term') currentSubst
+                    Map.map (Substitution.substituteInTerm $ Map.singleton var term') currentSubst
                 newSubst = Map.insert var term' currentSubst'
             modify $ \s -> s{mSubstitution = newSubst}
